--- a/ulid.go
+++ b/ulid.go
@@ -24,7 +24,6 @@ import (
 	"math/rand"
 	"sync"
 	"time"
-	"unsafe"
 )
 
 /*
@@ -493,10 +492,10 @@ func (id *ULID) SetEntropy(e []byte) error {
 // Compare returns an integer comparing id and other lexicographically.
 // The result will be 0 if id==other, -1 if id < other, and +1 if id > other.
 func (id ULID) Compare(other ULID) int {
-	ih := *(*uint64)(unsafe.Pointer(&id[0]))
-	il := *(*uint64)(unsafe.Pointer(&id[8]))
-	oh := *(*uint64)(unsafe.Pointer(&other[0]))
-	ol := *(*uint64)(unsafe.Pointer(&other[8]))
+	ih := binary.NativeEndian.Uint64(id[0:8])
+	il := binary.NativeEndian.Uint64(id[8:16])
+	oh := binary.NativeEndian.Uint64(other[0:8])
+	ol := binary.NativeEndian.Uint64(other[8:16])
 	if ih > oh {
 		return 1
 	}