This is the fix for https://github.com/oklog/ulid/pull/125

New:
```
goos: linux
goarch: amd64
pkg: github.com/oklog/ulid/v2
cpu: Intel(R) Xeon(R) Platinum 8370C CPU @ 2.80GHz
=== RUN   BenchmarkNewCompare
BenchmarkNewCompare
BenchmarkNewCompare-2           770801167                1.494 ns/op    21421.13 MB/s          0 B/op          0 allocs/op
PASS
```

Old:
```
goos: linux
goarch: amd64
pkg: github.com/oklog/ulid/v2
cpu: Intel(R) Xeon(R) Platinum 8370C CPU @ 2.80GHz
=== RUN   BenchmarkOldCompare
BenchmarkOldCompare
BenchmarkOldCompare-2           414179769                3.045 ns/op    10509.64 MB/s          0 B/op          0 allocs/op
PASS
```

It can be seen that the time for ULID Compare has decreased by about 51%! 

All test cases and benchmarks have passed!