@peterbourgon 2025-05-21
ULIDs are always big-endian, so if you want to look at a ULID as two hi/lo uint64s without copying, I think you can just do e.g.

```go
var id ULID
hi := binary.BigEndian.Uint64(id[:8])
lo := binary.BigEndian.Uint64(id[8:])
```

Extending that to a full Compare method might look like e.g.

```go
func (id ULID) Compare(other ULID) int {
	ahi := binary.BigEndian.Uint64(id[:8]) 
	bhi := binary.BigEndian.Uint64(other[:8])
	if ahi < bhi {
		return -1
	}
	if ahi > bhi {
		return 1
	}
	alo := binary.BigEndian.Uint64(id[8:]) 
	blo := binary.BigEndian.Uint64(other[8:])
	if alo < blo {
		return -1
	}
	if alo > blo {
		return 1
	}
	return 0
}
```

I think your proposed Compare methods actually return incorrect results for e.g. `01JVR914GPGV0BTVAK499X0591` vs `01JVR915GP0DP7QFH5JKQ1HETF`.

The current BenchmarkCompare uses the same two ULIDs in each b.N loop, which differ immediately in their first byte, so they're not really exercising the Compare method(s) in a useful way. One quick guess at a more useful benchmark would be e.g.

```go

func BenchmarkCompare(b *testing.B) {
	for _, testcase := range []struct {
		name string
		a    ulid.ULID
		b    ulid.ULID
	}{
		{
			name: "zeroes",
			a:    ulid.ULID{},
			b:    ulid.ULID{},
		},
		{
			name: "time a/b",
			a:    ulid.MustParseStrict(`01JVR914GPGV0BTVAK499X0591`),
			b:    ulid.MustParseStrict(`01JVR915GP0DP7QFH5JKQ1HETF`),
		},
		{
			name: "time b/a",
			a:    ulid.MustParseStrict(`01JVR915GP0DP7QFH5JKQ1HETF`),
			b:    ulid.MustParseStrict(`01JVR914GPGV0BTVAK499X0591`),
		},
		{
			name: "entropy by 1",
			a:    ulid.MustParseStrict(`01JVR915GP0DP7QFH5JKQ1HET0`),
			b:    ulid.MustParseStrict(`01JVR915GP0DP7QFH5JKQ1HET1`),
		},
		{
			name: "handcrafted",
			a:    ulid.ULID{5, 4, 3, 2, 1},
			b:    ulid.ULID{1, 2, 3, 4, 5},
		},
	} {
		b.Run(testcase.name, func(b *testing.B) {
			for _, fn := range []struct {
				name string
				fn   func(a, b ulid.ULID) int
			}{
				{"CompareBytes", ulid.ULID.CompareBytes},
				{"CompareUint64", ulid.ULID.Compare64},
			} {
				b.Run(fn.name, func(b *testing.B) {
					b.ReportAllocs()

					var want int
					switch {
					case testcase.a.String() < testcase.b.String():
						want = -1
					case testcase.a.String() > testcase.b.String():
						want = 1
					}

					for i := 0; i < b.N; i++ {
						if have := fn.fn(testcase.a, testcase.b); want != have {
							b.Fatalf("%s Compare %s: want %v, have %v", testcase.a, testcase.b, want, have)
						}
					}
				})
			}
		})
	}
}
```

which for my arm64 macbook pro returns

```

goos: darwin
goarch: arm64
pkg: github.com/oklog/ulid/v2
cpu: Apple M3 Pro
BenchmarkCompare
BenchmarkCompare/zeroes
BenchmarkCompare/zeroes/CompareBytes
BenchmarkCompare/zeroes/CompareBytes-11         	475559071	         2.423 ns/op	       0 B/op	       0 allocs/op
BenchmarkCompare/zeroes/CompareUint64
BenchmarkCompare/zeroes/CompareUint64-11        	740542461	         1.623 ns/op	       0 B/op	       0 allocs/op
BenchmarkCompare/time_a/b
BenchmarkCompare/time_a/b/CompareBytes
BenchmarkCompare/time_a/b/CompareBytes-11       	553243116	         2.165 ns/op	       0 B/op	       0 allocs/op
BenchmarkCompare/time_a/b/CompareUint64
BenchmarkCompare/time_a/b/CompareUint64-11      	873565628	         1.377 ns/op	       0 B/op	       0 allocs/op
BenchmarkCompare/time_b/a
BenchmarkCompare/time_b/a/CompareBytes
BenchmarkCompare/time_b/a/CompareBytes-11       	555538088	         2.167 ns/op	       0 B/op	       0 allocs/op
BenchmarkCompare/time_b/a/CompareUint64
BenchmarkCompare/time_b/a/CompareUint64-11      	869904038	         1.375 ns/op	       0 B/op	       0 allocs/op
BenchmarkCompare/entropy_by_1
BenchmarkCompare/entropy_by_1/CompareBytes
BenchmarkCompare/entropy_by_1/CompareBytes-11   	491751549	         2.437 ns/op	       0 B/op	       0 allocs/op
BenchmarkCompare/entropy_by_1/CompareUint64
BenchmarkCompare/entropy_by_1/CompareUint64-11  	735243434	         1.647 ns/op	       0 B/op	       0 allocs/op
BenchmarkCompare/handcrafted
BenchmarkCompare/handcrafted/CompareBytes
BenchmarkCompare/handcrafted/CompareBytes-11    	552388168	         2.164 ns/op	       0 B/op	       0 allocs/op
BenchmarkCompare/handcrafted/CompareUint64
BenchmarkCompare/handcrafted/CompareUint64-11   	868318581	         1.380 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/oklog/ulid/v2	14.032s
```

Of course the performance of CompareUint64 would be totally different on a 32-bit architecture, probably CompareBytes would be faster there!

This is just my first pass response. I'm not sure how I feel about architecture-specific optimizations in general, and I'm not sure how to evaluate architecture-specific optimizations that give ~2x improvement -- to me this is not a lot, if it were 10x or 100x that's a different thing!