@peterbourgon 2025-05-21
AFAIK one of the primary purposes of functions like `binary.BigEndian.Uint64` is to parse input bytes of a known endian-ness into a primitive value (in this case a uint64) that's correct for the current architecture, whatever it is. That is, for the same input bytes, it should produce a uint64 with little-endian byte order on a little-endian host, and the same uint64 with big-endian byte order on a big-endian host.

> However, running on a CPU in Big Endian mode, the code may behave abnormally (e.g. IBM POWER processors) because *uint64(*void)(&a[0]) will get the 0x01020304 result directly!

I would be very surprised if this were the case! Do you have any way to test this claim out?